Clearly define DTO and DAO
