+++ competition/models/membership.go
@@ -1,39 +1,20 @@
 package models
 
 import (
-	"database/sql/driver"
 	"fmt"
 
 	"gorm.io/gorm"
+
+	"arkavidia-backend-8.0/competition/types"
 )
-
-type MembershipRole string
-
-const (
-	Leader MembershipRole = "leader"
-	Member MembershipRole = "member"
-)
-
-func (membershipRole *MembershipRole) Scan(value interface{}) error {
-	*membershipRole = MembershipRole(value.(string))
-	return nil
-}
-
-func (membershipRole MembershipRole) Value() (driver.Value, error) {
-	return string(membershipRole), nil
-}
-
-func (MembershipRole) GormDataType() string {
-	return "membership_role"
-}
 
 type Membership struct {
 	gorm.Model
-	TeamID        uint           `json:"team_id" gorm:"uniqueIndex:membership_index"`
-	ParticipantID uint           `json:"participant_id" gorm:"uniqueIndex:membership_index"`
-	Role          MembershipRole `json:"role" gorm:"not null"`
-	Team          Team           `json:"team" gorm:"foreignKey:TeamID;references:ID"`
-	Participant   Participant    `json:"participant" gorm:"foreignKey:ParticipantID;references:ID"`
+	TeamID        uint                 `json:"team_id" gorm:"uniqueIndex:membership_index"`
+	ParticipantID uint                 `json:"participant_id" gorm:"uniqueIndex:membership_index"`
+	Role          types.MembershipRole `json:"role" gorm:"not null"`
+	Team          Team                 `json:"team" gorm:"foreignKey:TeamID;references:ID"`
+	Participant   Participant          `json:"participant" gorm:"foreignKey:ParticipantID;references:ID"`
 }
 
 // Menambahkan constraint untuk mengecek apakah terdapat participant yang mengikuti dua team atau lebih
@@ -50,7 +31,7 @@ func (membership *Membership) BeforeSave(tx *gorm.DB) error {
 			if oldMembership.Team.TeamCategory != "" && oldMembership.Team.TeamCategory == membership.Team.TeamCategory {
 				return fmt.Errorf("ERROR: CANNOT PARTICIPATE MORE THAN ONCE")
 			}
-			if oldMembership.Role == Leader && membership.Role == Leader {
+			if oldMembership.Role == types.Leader && membership.Role == types.Leader {
 				return fmt.Errorf("ERROR: INELIGIBLE LEADER")
 			}
 		}
@@ -62,7 +43,7 @@ func (membership *Membership) BeforeSave(tx *gorm.DB) error {
 		}
 
 		for _, oldMembership := range oldMembershipsTeamID {
-			if oldMembership.Role == Leader && membership.Role == Leader {
+			if oldMembership.Role == types.Leader && membership.Role == types.Leader {
 				return fmt.Errorf("ERROR: INELIGIBLE LEADER")
 			}
 		}
