+++ competition/models/team.go
@@ -1,76 +1,34 @@
 package models
 
 import (
-	"database/sql/driver"
 	"fmt"
 
 	"gorm.io/gorm"
+
+	"arkavidia-backend-8.0/competition/types"
 )
-
-type TeamCategory string
-
-const (
-	CP         TeamCategory = "competitive-programming"
-	Datavidia  TeamCategory = "datavidia"
-	UXVidia    TeamCategory = "uxvidia"
-	Arkalogica TeamCategory = "arkalogica"
-)
-
-func (teamCategory *TeamCategory) Scan(value interface{}) error {
-	*teamCategory = TeamCategory(value.(string))
-	return nil
-}
-
-func (teamCategory TeamCategory) Value() (driver.Value, error) {
-	return string(teamCategory), nil
-}
-
-func (TeamCategory) GormDataType() string {
-	return "team_category"
-}
-
-type TeamStatus string
-
-const (
-	WaitingForEvaluation TeamStatus = "waiting-for-evaluation"
-	Passed               TeamStatus = "passed"
-	Eliminated           TeamStatus = "eliminated"
-)
-
-func (teamStatus *TeamStatus) Scan(value interface{}) error {
-	*teamStatus = TeamStatus(value.(string))
-	return nil
-}
-
-func (teamStatus TeamStatus) Value() (driver.Value, error) {
-	return string(teamStatus), nil
-}
-
-func (TeamStatus) GormDataType() string {
-	return "team_status"
-}
 
 type Team struct {
 	gorm.Model
-	Username       string          `json:"username" gorm:"not null;unique"`
-	HashedPassword EncryptedString `json:"password" gorm:"not null" visibility:"false"`
-	TeamName       string          `json:"team_name" gorm:"not null;unique"`
-	TeamCategory   TeamCategory    `json:"team_category" gorm:"default:null"`
-	AdminID        uint            `json:"admin_id" gorm:"default:null"`
-	Status         TeamStatus      `json:"status" gorm:"not null"`
-	Memberships    []Membership    `json:"memberships"`
-	Submissions    []Submission    `json:"submissions"`
-	ApprovedBy     Admin           `json:"admin" gorm:"foreignKey:AdminID;references:ID"`
+	Username       string                `json:"username" gorm:"not null;unique"`
+	HashedPassword types.EncryptedString `json:"password" gorm:"not null" visibility:"false"`
+	TeamName       string                `json:"team_name" gorm:"not null;unique"`
+	TeamCategory   types.TeamCategory    `json:"team_category" gorm:"default:null"`
+	AdminID        uint                  `json:"admin_id" gorm:"default:null"`
+	Status         types.TeamStatus      `json:"status" gorm:"not null"`
+	Memberships    []Membership          `json:"memberships"`
+	Submissions    []Submission          `json:"submissions"`
+	ApprovedBy     Admin                 `json:"admin" gorm:"foreignKey:AdminID;references:ID"`
 }
 
 // Menambahkan constraint untuk mengecek apakah terdapat photos yang telah diapprove namun admin tidak tercatat
 // atau photos yang belum diapprove namun admin tercatat
 func (team *Team) BeforeSave(tx *gorm.DB) error {
 	if team.Status != "" {
-		if team.Status != WaitingForEvaluation && team.AdminID == 0 {
+		if team.Status != types.WaitingForEvaluation && team.AdminID == 0 {
 			return fmt.Errorf("ERROR: ADMIN MUST BE RECORDED")
 		}
-		if team.Status == WaitingForEvaluation && team.AdminID != 0 {
+		if team.Status == types.WaitingForEvaluation && team.AdminID != 0 {
 			return fmt.Errorf("ERROR: STATUS MUST BE RECORDED")
 		}
 	}
