+++ competition/repository/admin.go
@@ -1,0 +1,6 @@
+package repository
+
+type SignInAdminRequest struct {
+	Username string `json:"username" binding:"required,alphanum"`
+	Password string `json:"password" binding:"required,ascii"`
+}
