+++ competition/services/database/plugins.go
@@ -1,6 +1,6 @@
 package database
 
-// TODO: Tambahkan gorm plugin menggunakan go-migration untuk menginisialisasikan type pada PosgreSQL
+// TODO: Tambahkan gorm plugin menggunakan go-migration untuk menginisialisasikan type pada PostgreSQL
 // REFERENCE: https://gorm.io/docs/write_plugins.html
 // REFERENCE: https://github.com/golang-migrate/migrate#use-in-your-go-project
 // ASSIGNED TO: @rayhankinan
