+++ competition/types/encrypted-string.go
@@ -1,4 +1,4 @@
-package models
+package types
 
 import (
 	"database/sql/driver"
